# internal/api/swagger/swagger.yaml
basePath: /v1
definitions:
...
  internal_api_handler.envelope:
    properties:
      data: {}
      error:
        type: string
    type: object
  internal_api_handler.giftRequest:
    properties:
      amount_cents:
        type: integer
      currency:
        type: string
      date:
        type: string
      debt_type:
        type: string
      direction:
        type: string
      notes:
        type: string
      person_id:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: kith
    url: https://github.com/nhymxu/kith-pms
  description: Self-hosted personal relationship manager API.
  license:
    name: MIT
  title: kith API
  version: "1.0"
paths:
...
  /gifts:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - description: Filter by person ID
        in: query
        name: person_id
        type: integer
      - description: 'Filter: planned, given, received'
        in: query
        name: direction
        type: string
      - description: 'Filter: i_owe, they_owe'
        in: query
        name: debt_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: List gifts
      tags:
      - gifts
    post:
      consumes:
      - application/json
      parameters:
      - description: Gift data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.giftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Create gift
      tags:
      - gifts
  /gifts/{id}:
    delete:
      parameters:
      - description: Gift ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Delete gift
      tags:
      - gifts
    get:
      parameters:
      - description: Gift ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Get gift
      tags:
      - gifts
    put:
      consumes:
      - application/json
      parameters:
      - description: Gift ID
        in: path
        name: id
        required: true
        type: integer
      - description: Gift data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.giftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Update gift
      tags:
      - gifts
  /gifts/{id}/image:
    delete:
      parameters:
      - description: Gift ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Delete gift image
      tags:
      - gifts
    get:
      parameters:
      - description: Gift ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Get gift image
      tags:
      - gifts
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Gift ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file (jpeg/png/gif/webp, max 5MB)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    uploaded:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Upload gift image
      tags:
      - gifts
...
  /people/{id}/gifts:
    get:
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: List gifts for person
      tags:
      - gifts
...
securityDefinitions:
  CSRFHeader:
    description: Must be set to "kith-spa" on all mutating requests (POST/PUT/PATCH/DELETE)
    in: header
    name: X-Requested-With
    type: apiKey
  CookieAuth:
    description: HttpOnly session cookie set by POST /v1/auth/login
    in: cookie
    name: kith_session
    type: apiKey
swagger: "2.0"